Magnesium

Documentation for Magnesium.

Magnesium.Concepts.ControllerComponent

Magnesium.Concepts.InZoneComponent

Magnesium.Concepts.IsZoneComponent

Magnesium.Concepts.OwnerComponent

Magnesium.Concepts.PlayerComponent

Magnesium.Concepts.ShuffleComponent

Magnesium.Concepts.ShuffleDeckComponent

Magnesium.Concepts.specify

Magnesium.Concepts.specify — Method

specify(io::IO, e::Entity, m::AbstractLedger)

Print a description that specifies e to io.

Magnesium.Concepts.ControllerComponent — Type

An Entity with a ControllerComponent has a controller.

Magnesium.Concepts.InZoneComponent — Type

An Entity with an InZoneComponent component is in a zone.

Magnesium.Concepts.IsZoneComponent — Type

An Entity with a IsZoneComponent component is a zone entity.

Example

# this zone is Entity(1)'s hand
IsZoneComponent{OwnedZone}(Hand, Entity(1))

# this zone is the stack
IsZoneComponent{UnownedZone}(Stack, nothing)

Magnesium.Concepts.OwnerComponent — Type

An Entity with an OwnerComponent has an owner.

Magnesium.Concepts.PlayerComponent — Type

An Entity with a PlayerComponent is a player.

Magnesium.Concepts.ShuffleComponent — Type

Denotes that a zone entity needs the top or bottom n cards shuffled.

Magnesium.Concepts.ShuffleDeckComponent — Type

Denotes that a zone entity needs to be shuffled.